chore(upgrade): separate upgrade package and build tag
On-behalf-of: @SAP christopher.junk@sap.com
